Integrate the new Endpoint class in 3.0
This fixes the configuration not actually being utilized (as it is no
longer supposed to be passed to the adapter object but rather to a
new Endpoint object). This integration is by no means optimal but it
works and preserves backwards compability. In the future the configuration
should probably follow the Solarium model more closely, but that'd break
backwards compability.

Read more about it here:

http://wiki.solarium-project.org/index.php/V3:Upgrading_from_Solarium_2.x#Endpoints